refactored provider classes, added javadoc with intented usages
