+++ src/main/java/io/github/mzmine/gui/chartbasics/simplechart/providers/impl/spectra/CachedMobilityScan.java
@@ -16,7 +16,7 @@
  *  USA
  */
 
-package io.github.mzmine.modules.visualization.rawdataoverviewims.providers;
+package io.github.mzmine.gui.chartbasics.simplechart.providers.impl.spectra;
 
 import com.google.common.collect.Range;
 import io.github.mzmine.datamodel.DataPoint;
@@ -26,17 +26,12 @@ import io.github.mzmine.datamodel.MassList;
 import io.github.mzmine.datamodel.MassSpectrumType;
 import io.github.mzmine.datamodel.MobilityScan;
 import io.github.mzmine.datamodel.MobilityType;
-import io.github.mzmine.datamodel.PolarityType;
 import io.github.mzmine.datamodel.RawDataFile;
-import io.github.mzmine.datamodel.impl.SimpleFrame;
 import io.github.mzmine.util.DataPointUtils;
 import java.nio.DoubleBuffer;
-import java.util.ArrayList;
-import java.util.Comparator;
+import java.util.Arrays;
 import java.util.Iterator;
-import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -46,107 +41,22 @@ import javax.annotation.Nullable;
  *
  * @author https://github.com/SteffenHeu
  */
-public class CachedFrame implements Frame {
+public class CachedMobilityScan implements MobilityScan {
 
+  private final MobilityScan originalMobilityScan;
   private final double[] mzs;
   private final double[] intensities;
-  private final Frame originalFrame;
-  private List<MobilityScan> mobilityScans;
-  private List<MobilityScan> sortedMobilityScans;
+  private final double tic;
 
-  public CachedFrame(SimpleFrame frame, double frameNoiseLevel, double mobilityScaNoiseLevel) {
-    originalFrame = frame;
+  public CachedMobilityScan(MobilityScan scan, double noiseLevel) {
+    this.originalMobilityScan = scan;
 
     double[][] data = DataPointUtils
-        .getDatapointsAboveNoiseLevel(frame.getMzValues(), frame.getIntensityValues(),
-            frameNoiseLevel);
+        .getDatapointsAboveNoiseLevel(scan.getMzValues(), scan.getIntensityValues(), noiseLevel);
 
     mzs = data[0];
     intensities = data[1];
-
-    this.mobilityScans = new ArrayList<>();
-    for (MobilityScan scan : frame.getMobilityScans()) {
-      mobilityScans.add(new CachedMobilityScan(scan, mobilityScaNoiseLevel));
-    }
-    sortedMobilityScans = mobilityScans.stream()
-        .sorted(Comparator.comparingDouble(MobilityScan::getMobility)).collect(Collectors.toList());
-  }
-
-
-  public double[] getIntensities() {
-    return intensities;
-  }
-
-  public double[] getMzs() {
-    return mzs;
-  }
-
-  @Override
-  public int getNumberOfMobilityScans() {
-    return mobilityScans.size();
-  }
-
-  @Nonnull
-  @Override
-  public MobilityType getMobilityType() {
-    return originalFrame.getMobilityType();
-  }
-
-  @Nonnull
-  @Override
-  public Range<Double> getMobilityRange() {
-    return originalFrame.getMobilityRange();
-  }
-
-  @Nullable
-  @Override
-  public MobilityScan getMobilityScan(int num) {
-    return mobilityScans.get(num);
-  }
-
-  @Nonnull
-  @Override
-  public List<MobilityScan> getMobilityScans() {
-    return mobilityScans;
-  }
-
-  @Nonnull
-  @Override
-  public List<MobilityScan> getSortedMobilityScans() {
-    return sortedMobilityScans;
-  }
-
-  @Override
-  public double getMobilityForMobilityScanNumber(int mobilityScanIndex) {
-    return originalFrame.getMobilityForMobilityScanNumber(mobilityScanIndex);
-  }
-
-  @Override
-  public double getMobilityForMobilityScan(MobilityScan scan) {
-    return originalFrame.getMobilityForMobilityScan(scan);
-  }
-
-  @Nullable
-  @Override
-  public DoubleBuffer getMobilities() {
-    return originalFrame.getMobilities();
-  }
-
-  @Nonnull
-  @Override
-  public Set<ImsMsMsInfo> getImsMsMsInfos() {
-    return originalFrame.getImsMsMsInfos();
-  }
-
-  @Nullable
-  @Override
-  public ImsMsMsInfo getImsMsMsInfoForMobilityScan(int mobilityScanNumber) {
-    return originalFrame.getImsMsMsInfoForMobilityScan(mobilityScanNumber);
-  }
-
-  @Override
-  public void addMobilityScan(MobilityScan originalMobilityScan) {
-    originalFrame.addMobilityScan(originalMobilityScan);
+    tic = Arrays.stream(intensities).sum();
   }
 
   @Override
@@ -156,7 +66,7 @@ public class CachedFrame implements Frame {
 
   @Override
   public MassSpectrumType getSpectrumType() {
-    return originalFrame.getSpectrumType();
+    return originalMobilityScan.getSpectrumType();
   }
 
   @Nonnull
@@ -186,13 +96,13 @@ public class CachedFrame implements Frame {
   @Nullable
   @Override
   public Double getBasePeakMz() {
-    return originalFrame.getBasePeakMz();
+    return originalMobilityScan.getBasePeakMz();
   }
 
   @Nullable
   @Override
   public Double getBasePeakIntensity() {
-    return originalFrame.getBasePeakIntensity();
+    return originalMobilityScan.getBasePeakIntensity();
   }
 
   @Nullable
@@ -205,15 +115,13 @@ public class CachedFrame implements Frame {
   @Nullable
   @Override
   public Range<Double> getDataPointMZRange() {
-    throw new UnsupportedOperationException(
-        "Not intended. This frame is used for visualisation only");
+    return originalMobilityScan.getDataPointMZRange();
   }
 
   @Nullable
   @Override
   public Double getTIC() {
-    throw new UnsupportedOperationException(
-        "Not intended. This frame is used for visualisation only");
+    return tic;
   }
 
   @Override
@@ -225,52 +133,37 @@ public class CachedFrame implements Frame {
   @Nonnull
   @Override
   public RawDataFile getDataFile() {
-    return originalFrame.getDataFile();
+    return originalMobilityScan.getDataFile();
   }
 
   @Override
-  public int getScanNumber() {
-    return originalFrame.getScanNumber();
-  }
-
-  @Nonnull
-  @Override
-  public String getScanDefinition() {
-    return originalFrame.getScanDefinition();
+  public double getMobility() {
+    return originalMobilityScan.getMobility();
   }
 
   @Override
-  public int getMSLevel() {
-    return originalFrame.getMSLevel();
+  public MobilityType getMobilityType() {
+    return originalMobilityScan.getMobilityType();
+  }
+
+  @Override
+  public Frame getFrame() {
+    return originalMobilityScan.getFrame();
   }
 
   @Override
   public float getRetentionTime() {
-    return originalFrame.getRetentionTime();
+    return originalMobilityScan.getRetentionTime();
   }
 
-  @Nonnull
   @Override
-  public Range<Double> getScanningMZRange() {
-    return originalFrame.getScanningMZRange();
-  }
-
-  @Nonnull
-  @Override
-  public PolarityType getPolarity() {
-    return originalFrame.getPolarity();
-  }
-
-  @Nonnull
-  @Override
-  public MassList[] getMassLists() {
-    throw new UnsupportedOperationException(
-        "Not intended. This frame is used for visualisation only");
+  public int getMobilityScamNumber() {
+    return originalMobilityScan.getMobilityScamNumber();
   }
 
   @Nullable
   @Override
-  public MassList getMassList(@Nonnull String name) {
+  public ImsMsMsInfo getMsMsInfo() {
     throw new UnsupportedOperationException(
         "Not intended. This frame is used for visualisation only");
   }
@@ -283,6 +176,19 @@ public class CachedFrame implements Frame {
 
   @Override
   public void removeMassList(@Nonnull MassList massList) {
+    throw new UnsupportedOperationException(
+        "Not intended. This frame is used for visualisation only");
+  }
+
+  @Nonnull
+  @Override
+  public Set<MassList> getMassLists() {
+    throw new UnsupportedOperationException(
+        "Not intended. This frame is used for visualisation only");
+  }
+
+  @Override
+  public MassList getMassList(@Nonnull String name) {
     throw new UnsupportedOperationException(
         "Not intended. This frame is used for visualisation only");
   }
